Move BSV runloop code over to runloop.c
